+++ Sources/Swift/Integrations/UserFeedback/SentryUserFeedbackWidgetButtonView.swift
@@ -1,3 +1,4 @@
+#if !SDK_V10
 import Foundation
 #if os(iOS) && !SENTRY_NO_UI_FRAMEWORK
 internal import _SentryPrivate
@@ -192,3 +193,4 @@ final class SentryUserFeedbackWidgetButtonView: UIView {
 }
 
 #endif // os(iOS) && !SENTRY_NO_UI_FRAMEWORK
+#endif // !SDK_V10