+++ Tests/SentryTests/HybridSDK/SentryInternalReplayApiIntegrationTests.swift
@@ -10,6 +10,7 @@ class SentryInternalReplayApiIntegrationTests: XCTestCase {
     private static let validReplayId = "0eac7ab503354dd5819b03e263627a29"
 
     override func tearDown() {
+        // swiftlint:disable:next avoid_clear_test_state - just disabled to allow adding the SwiftLint rule. Please double check if you can remove this when touching this.
         clearTestState()
         super.tearDown()
     }