+++ Tests/SentryTests/HybridSDK/SentryInternalViewHierarchyApiIntegrationTests.swift
@@ -17,6 +17,7 @@ class SentryInternalViewHierarchyApiIntegrationTests: XCTestCase {
     }
 
     override func tearDown() {
+        // swiftlint:disable:next avoid_clear_test_state - just disabled to allow adding the SwiftLint rule. Please double check if you can remove this when touching this.
         clearTestState()
         super.tearDown()
     }