+++ Tests/SentryTests/Integrations/Performance/CoreData/SentryCoreDataTrackingIntegrationTest.swift
@@ -43,6 +43,7 @@ class SentryCoreDataTrackingIntegrationTests: XCTestCase {
     override func tearDownWithError() throws {
         super.tearDown()
         try fixture.coreDataStack.reset()
+        // swiftlint:disable:next avoid_clear_test_state - just disabled to allow adding the SwiftLint rule. Please double check if you can remove this when touching this.
         clearTestState()
     }