+++ Tests/SentryTests/ViewCapture/SentryRedactViewHelperTests.swift
@@ -8,6 +8,7 @@ import XCTest
 final class SentryRedactViewHelperTests: XCTestCase {
 
     override func tearDown() {
+        // swiftlint:disable:next avoid_clear_test_state - just disabled to allow adding the SwiftLint rule. Please double check if you can remove this when touching this.
         clearTestState()
         super.tearDown()
     }